+++ demos/object_detection/include/yolo.h
@@ -10,6 +10,7 @@
 
 struct YoloLayerParam
 {
+    std::string name;
     int numGridX;
     int numGridY;
     int numBoxes;
@@ -19,10 +20,10 @@ struct YoloLayerParam
     float scaleX=0;
     float scaleY=0;
     YoloLayerParam() = default;
-    YoloLayerParam(int _gx, int _gy, int _numB, 
+    YoloLayerParam(std::string _name, int _gx, int _gy, int _numB, 
                    const std::vector<float> &_vAnchorW, const std::vector<float> &_vAnchorH, const std::vector<int> &_vTensorIdx, 
                    float _sx = 0.f, float _sy = 0.f)
-    :numGridX(_gx), numGridY(_gy), numBoxes(_numB), 
+    :name(_name), numGridX(_gx), numGridY(_gy), numBoxes(_numB), 
      anchorWidth(_vAnchorW), anchorHeight(_vAnchorH),
      tensorIdx(_vTensorIdx), scaleX(_sx), scaleY(_sy)
     {}
@@ -60,6 +61,7 @@ public:
     ~Yolo();
     Yolo();
     Yolo(YoloParam &_cfg);
+    void LayerReorder(dxrt::Tensors output_info);
     void LayerInverse(int mode);
     /* for concated tensor */
     std::vector< BoundingBox > PostProc(float *data);