+++ demos/object_detection/yolo_demo_l1.cpp
@@ -40,16 +40,15 @@ using namespace cv;
 #define DMA_IOCTL_DATA		_IOW('T', 3, struct npu_dma_ioctl_data_copy)
 
 // pre/post parameter table
-extern YoloParam yolov5s_320, yolov5s_512, yolov5s_640, yolov5s_512_concat, yolox_s_512, yolov7_640, yolov7_512, yolov4_608;
+extern YoloParam yolov5s_320, yolov5s_512, yolov5s_640, yolox_s_512, yolov7_640, yolov7_512, yolov4_608;
 YoloParam yoloParams[] = {
     [0] = yolov5s_320,
     [1] = yolov5s_512,
     [2] = yolov5s_640,
-    [3] = yolov5s_512_concat,
-    [4] = yolox_s_512,
-    [5] = yolov7_640,
-    [6] = yolov7_512,
-    [7] = yolov4_608
+    [3] = yolox_s_512,
+    [4] = yolov7_640,
+    [5] = yolov7_512,
+    [6] = yolov4_608
 };
 
 
@@ -273,6 +272,7 @@ int main(int argc, char *argv[])
     auto ie = dxrt::InferenceEngine(modelPath, &option);
     auto yoloParam = yoloParams[paramIdx];
     Yolo yolo = Yolo(yoloParam);
+    yolo.LayerReorder(ie.outputs());
     auto& profiler = dxrt::Profiler::GetInstance();
     dxrt::CleanMemIf(); // cache flush for riscv64
     if(!imgFile.empty())